refactor: use atomic design pattern
