+++ components/atoms/TwitterIcon.tsx
@@ -1,4 +1,4 @@
-export default function ({ className }: { className?: string }) {
+const TwitterIcon = ({ className }: { className?: string }) => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -10,3 +10,4 @@ export default function ({ className }: { className?: string }) {
     </svg>
   )
 }
+export default TwitterIcon
