+++ components/index.tsx
@@ -1,0 +1,8 @@
+export { default as Layout } from './templates/Layout'
+export { default as GradientBackground } from './atoms/GradientBackground'
+export { default as Navigation } from './organisms/Navigation'
+export { default as Posts } from './molecules/Posts'
+export { default as ProfileImage } from './atoms/ProfileImage'
+export { default as TwitterIcon } from './atoms/TwitterIcon'
+export { default as ContentLink } from './atoms/ContentLink'
+export { default as Footer } from './organisms/Footer'
