--- layouts/index.tsx
@@ -1,8 +1,0 @@
-export { default as Layout } from './Layout'
-export { default as GradientBackground } from './GradientBackground'
-export { default as Navigation } from './Navigation'
-export { default as Posts } from './Posts'
-export { default as ProfileImage } from './ProfileImage'
-export { default as TwitterIcon } from './TwitterIcon'
-export { ContentLink } from './ContentLink'
-export { Footer } from './Footer'
