+++ tsconfig.json
@@ -17,7 +17,7 @@
     "types": ["node"],
     "baseUrl": ".",
     "paths": {
-      "@/layouts/*": ["layouts/*"],
+      "@/components/*": ["components/*"],
       "@/lib/*": ["lib/*"],
       "@/public/*": ["public/*"]
     },
